create factory folder, move user creation in factory
