+++ backend/test/helpers.ts
@@ -4,9 +4,6 @@
 import { createTestClient } from 'apollo-server-testing'
 import createServer from '../src/server/createServer'
 import { initialize } from '@dbTools/helpers'
-import { createUser, setPassword } from '@/seeds/graphql/mutations'
-import { LoginEmailOptIn } from '@entity/LoginEmailOptIn'
-import { User } from '@entity/User'
 import { entities } from '@entity/index'
 
 export const headerPushMock = jest.fn((t) => {
@@ -44,19 +41,6 @@ export const resetEntity = async (entity: any) => {
     const ids = items.map((i: any) => i.id)
     await entity.delete(ids)
   }
-}
-
-export const createConfirmedUser = async (mutate: any, user: any) => {
-  // resetToken()
-  await mutate({ mutation: createUser, variables: user })
-  const dbUser = await User.findOne({ where: { email: user.email } })
-  if (!dbUser) throw new Error('Ups, no user found')
-  const optin = await LoginEmailOptIn.findOne({ where: { userId: dbUser.id } })
-  if (!optin) throw new Error('Ups, no optin found')
-  await mutate({
-    mutation: setPassword,
-    variables: { password: 'Aa12345_', code: optin.verificationCode },
-  })
 }
 
 export const resetToken = () => {
